+++ integration-test/llm_mock/openai_chat.clj
@@ -5,6 +5,11 @@
    [llm-mock.mocks :as llm.mocks]
    [org.httpkit.server :as hk]))
 
+(def ^:dynamic *thinking-tag* "think")
+
+(defn set-thinking-tag! [tag]
+  (alter-var-root #'*thinking-tag* (constantly tag)))
+
 (defn ^:private send-sse!
   "Send a single SSE data line with a JSON payload, followed by a blank line."
   [ch m]
@@ -62,6 +67,28 @@
   (send-sse! ch {:choices [{:delta {} :finish_reason "stop"}]})
   (hk/close ch))
 
+(defn ^:private reasoning-text-0 [ch]
+  (send-sse! ch {:choices [{:delta {:content (str "<" *thinking-tag* ">")}}]})
+  (send-sse! ch {:choices [{:delta {:content "I should say"}}]})
+  (send-sse! ch {:choices [{:delta {:content " hello"}}]})
+  (send-sse! ch {:choices [{:delta {:content (str "</" *thinking-tag* ">")}}]})
+  (send-sse! ch {:choices [{:delta {:content "hello"}}]})
+  (send-sse! ch {:choices [{:delta {:content " there!"}}]})
+  (send-sse! ch {:usage {:prompt_tokens 10 :completion_tokens 20}})
+  (send-sse! ch {:choices [{:delta {} :finish_reason "stop"}]})
+  (hk/close ch))
+
+(defn ^:private reasoning-text-1 [ch]
+  (send-sse! ch {:choices [{:delta {:content (str "<" *thinking-tag* ">")}}]})
+  (send-sse! ch {:choices [{:delta {:content "I should say"}}]})
+  (send-sse! ch {:choices [{:delta {:content " fine"}}]})
+  (send-sse! ch {:choices [{:delta {:content (str "</" *thinking-tag* ">")}}]})
+  (send-sse! ch {:choices [{:delta {:content "I'm "}}]})
+  (send-sse! ch {:choices [{:delta {:content " fine"}}]})
+  (send-sse! ch {:usage {:prompt_tokens 10 :completion_tokens 20}})
+  (send-sse! ch {:choices [{:delta {} :finish_reason "stop"}]})
+  (hk/close ch))
+
 (defn ^:private chat-title-text-0 [ch]
   (send-sse! ch {:choices [{:delta {:content "Some Cool"}}]})
   (send-sse! ch {:choices [{:delta {:content " Title"}}]})
@@ -94,6 +121,8 @@
                        :simple-text-0 (simple-text-0 ch)
                        :simple-text-1 (simple-text-1 ch)
                        :simple-text-2 (simple-text-2 ch)
+                       :reasoning-0 (reasoning-text-0 ch)
+                       :reasoning-1 (reasoning-text-1 ch)
                        ;; default fallback
                        (do
                          (send-sse! ch {:choices [{:delta {:content "hello"}}]})